clustering-backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
scikit-learn==1.6.1